requirements.txt
Flask~=3.1.2
Flask-SQLAlchemy
SQLAlchemy~=2.0.44
python-dotenv~=1.2.1
Werkzeug~=3.1.3
# connections reseau et bluetooth
nmcli
pydbus~=0.6.0
mutagen~=1.47.0
pillow~=12.0.0
# flux rss pour podcasts
requests~=2.32.5
bs4~=0.0.2
feedparser~=6.0.12
